Add `CONNECTION_CLOSE` frame to h3i actions  (#1835)
* add connection_close frame to h3i

* Reviewer comments

* Refactor error prompts to new file
