+++ h3i/src/prompts/h3/mod.rs
@@ -33,6 +33,8 @@ use inquire::validator::Validation;
 use inquire::InquireError;
 use inquire::Select;
 use inquire::Text;
+use qlog::events::quic::ErrorSpace;
+use quiche::ConnectionError;
 
 use crate::actions::h3::Action;
 use crate::config::Config;
@@ -89,6 +91,8 @@ const EXTENSION: &str = "extension_frame";
 const OPEN_UNI_STREAM: &str = "open_uni_stream";
 const RESET_STREAM: &str = "reset_stream";
 const STOP_SENDING: &str = "stop_sending";
+const CONNECTION_CLOSE: &str = "connection_close";
+
 const COMMIT: &str = "commit";
 const FLUSH_PACKETS: &str = "flush_packets";
 const WAIT: &str = "wait";
@@ -153,6 +157,7 @@ impl Prompter {
             CANCEL_PUSH => prompt_cancel_push(),
             PUSH_PROMISE => prompt_push_promise(),
             PRIORITY_UPDATE => priority::prompt_priority(),
+            CONNECTION_CLOSE => prompt_connection_close(),
             FLUSH_PACKETS => return PromptOutcome::Action(Action::FlushPackets),
             COMMIT => return PromptOutcome::Commit,
             WAIT => prompt_wait(),
@@ -232,6 +237,7 @@ fn prompt_action() -> InquireResult<String> {
 }
 
 fn action_suggester(val: &str) -> SuggestionResult<Vec<String>> {
+    // TODO: make this an enum to automatically pick up new actions
     let suggestions = [
         HEADERS,
         HEADERS_NO_PSEUDO,
@@ -247,6 +253,7 @@ fn action_suggester(val: &str) -> SuggestionResult<Vec<String>> {
         OPEN_UNI_STREAM,
         RESET_STREAM,
         STOP_SENDING,
+        CONNECTION_CLOSE,
         FLUSH_PACKETS,
         COMMIT,
         WAIT,
@@ -423,6 +430,22 @@ fn prompt_extension() -> InquireResult<Action> {
     Ok(action)
 }
 
+pub fn prompt_connection_close() -> InquireResult<Action> {
+    let (error_space, error_code) = errors::prompt_transport_or_app_error()?;
+    let reason = Text::new("reason phrase:")
+        .with_placeholder("optional reason phrase")
+        .prompt()
+        .unwrap_or_default();
+
+    Ok(Action::ConnectionClose {
+        error: ConnectionError {
+            is_app: matches!(error_space, ErrorSpace::ApplicationError),
+            error_code,
+            reason: reason.as_bytes().to_vec(),
+        },
+    })
+}
+
 fn validate_wait_period(period: &str) -> SuggestionResult<Validation> {
     let x = period.parse::<u64>();
 
@@ -450,6 +473,7 @@ fn prompt_yes_no(msg: &str) -> InquireResult<bool> {
     Ok(res == YES)
 }
 
+mod errors;
 mod headers;
 mod priority;
 mod settings;
